requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg[binary]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
python-dotenv
jinja2
aiofiles
email-validator
pandas
openpyxl